Add getName
I created getName from the code that fails when compiling ponyc with
config=Debug to help debug issue #2776 from ponylang/ponyc:

  https://github.com/ponylang/ponyc/issues/2776

From my analysis the problem is that Expected::~Expected destructor
has two verions. One with NDEBUG is defined when compiling llvm in
Release mode and a different version when NDEBUG is not define when
compiling ponyc and its tests in Debug mode. Specifically in Debug mode
an additional class variable, Checked, is added and if Checked is false
when assertIsChecked called the program aborts, which is what I
reported in #2776.

The reason this occurs is because the destructor, ~Expected, is not inlined
in RuntimeDyld::loadObjectImpl and the linker called the Debug version
of ~Expected which calls the Debug version of assertIsChecked. Since
the the other methods, including the constructor, were inlined Checked is
never allocated nor updated. As a consequence assertIsChecked accesses
uninitialized memory which happens to be zero in my tests, and fails.